R6 Class representing a set of Parameters

R6 Class representing a set of Parameters

R6 Class representing a set of Parameters

Format

R6Class object.

Value

Object of R6Class with all the parameters related to the population and the epidemic.

Details

Parameters contains all the parameters related to the epidemic

Public fields

nage

number of age groups

preImmune

the proportion of immune individuals at the beginning of the epidemic

preExposed

the proportion of exposed individuals at the beginning of the epidemic

preInfected

the proportion of exposed individuals at the beginning of the epidemic

symptomatic

the proportion of symptomatic infection

progression

the rate of transfer from E to I

beta

the transmission probability upon contact with an infected (retro computed from R0)

contact

the contact matrix

agegroupnames

the name of the agegroups

Active bindings

susceptibility

A vector of size *nage* for susceptibility adjustment

R0

the basic reproduction number

removal

the rate of transfer from I to R

duration

the duration fo the simulation i.e. c("Week", "Month", "Trimester", "Semester", "Year").

nbDays

the number of days of the simulation

Methods

Public methods

Parameters$new()

Parameters$getList()

Parameters$clone()

Method new()

Create a new `Parameters` object.

Usage

Parameters$new(R0 = 2.2)

Arguments

R0

R0 (=3)

Returns

A new `Parameters` object.

Method getList()

Create a new list containing the required parameters for the model

Usage

Parameters$getList()

Returns

A new list of parameters.

Method clone()

The objects of this class are cloneable with this method.

Usage

Parameters$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

params = Parameters$new()